team(charity-fundraiser): add new team
The Russian aggression must be met with rapid answer. The charity organisations need funding, Ukraine suffers. Yes, people donate money already, but donating in general could be made fancier. NFT (and all other kinds of art or products too) is being sold anyways despite the conflict. Apparently there is no website of this sort yet on the market. It would be a great way for junior or mid developers to include a great global-scale initiative in their portfolio. And each team member would be of course a stakeholder once the start-up project is up and working. Come, join and lets do it! And also we could then use it, or the ideas or the knowledge for the upcoming Google Solutions Challenge, hitting multiple critical points out of the 17.